almost finished

+++ Flowers/ViewModel/AdminPanelViewModel.cs
@@ -1,7 +1,10 @@
-﻿using CommunityToolkit.Mvvm.ComponentModel;
+﻿using CommunityToolkit.Maui.Alerts;
+using CommunityToolkit.Maui.Core;
+using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 using Flowers.Models;
 using Flowers.Services;
+using Flowers.Views;
 using System.Collections.ObjectModel;
 using System.Threading.Tasks;
 
@@ -15,11 +18,18 @@ namespace Flowers.ViewModel
         [ObservableProperty] private ObservableCollection<User> users;
         [ObservableProperty] private ObservableCollection<Models.Flowers> flowers;
         [ObservableProperty] private ObservableCollection<Order> orders;
+        [ObservableProperty] private ObservableCollection<Store> stores;
 
+        [ObservableProperty] private bool isAddingUser = false;
+        [ObservableProperty] private User userToAdd;
+
+        [ObservableProperty] private Order selectedOrder;
+        [ObservableProperty] private bool isEditingOrder = false;
         public AdminPanelViewModel(DatabaseService databaseService)
         {
             _databaseService = databaseService;
             _databaseService.DataUpdated += OnDataUpdated;
+            UserToAdd = new User();
             LoadData();
         }
         private async void OnDataUpdated(object sender, EventArgs e)
@@ -33,6 +43,7 @@ namespace Flowers.ViewModel
             Users = new ObservableCollection<User>(await _databaseService.GetAllUsersAsync());
             Flowers = new ObservableCollection<Models.Flowers>(await _databaseService.GetAllFlowersAsync());
             Orders = new ObservableCollection<Order>(await _databaseService.GetAllOrdersAsync());
+            Stores = new ObservableCollection<Store>(await _databaseService.GetAllStoresAsync());
         }
 
         [RelayCommand]
@@ -45,23 +56,51 @@ namespace Flowers.ViewModel
         }
 
         [RelayCommand]
+        private void CloseModals() { 
+            IsAddingUser = false;
+            IsEditingOrder = false;
+        }
+
+        [RelayCommand]
         private async Task AddBouquet()
         {
-            // Логика для добавления букета
+            await Shell.Current.GoToAsync($"bouquetediting", true, new Dictionary<string, object> {
+                        {"Bouquet", new Bouquet()}
+                    });
         }
+
+        [RelayCommand]
+        private async Task EditBouquet(Bouquet bouquet)
+        {
+            await Shell.Current.GoToAsync($"bouquetediting", true, new Dictionary<string, object> {
+                        {"Bouquet", bouquet}
+                    });
+        }
+
 
         [RelayCommand]
         private async Task DeleteBouquet(Bouquet bouquet)
         {
-            // Логика для удаления букета
+            if (bouquet != null)
+            {
+                Bouquets.Remove(bouquet);
+                await _databaseService.AddBouquetAsync(bouquet);
+            }
         }
 
         [RelayCommand]
-        private async Task AddUser()
+        private async Task AddUserFromAdmin()
         {
-            // Логика для добавления пользователя
+            IsAddingUser = true;
         }
+        [RelayCommand]
+        private async Task AddUser() {
+            UserToAdd.Password = Encoder.ComputeHash(UserToAdd.Password, "SHA512", null);
+            await _databaseService.SaveUserAsync(UserToAdd);
 
+            await Toast.Make("Пользователь с именем '" + UserToAdd.Username + "' добавлен", ToastDuration.Short, 16).Show();
+            LoadData();
+        }
         [RelayCommand]
         private async Task DeleteUser(User user)
         {
@@ -73,36 +112,48 @@ namespace Flowers.ViewModel
         }
 
         [RelayCommand]
-        private async Task AddFlower()
+        private async Task AddFlower(Models.Flowers flower)
         {
-            // Логика для добавления цветка
+            if (flower != null)
+            {
+                Flowers.Remove(flower);
+                await _databaseService.AddFlowerAsync(flower);
+            }
         }
 
         [RelayCommand]
         private async Task DeleteFlower(Models.Flowers flower)
         {
-            /*if (flower != null)
+            if (flower != null)
             {
                 Flowers.Remove(flower);
                 await _databaseService.DeleteFlowerAsync(flower);
-            }*/
-        }
-
-        [RelayCommand]
-        private async Task ChangeOrderStatus(Order order)
-        {
-            // Логика для изменения статуса заказа
+            }
         }
 
         [RelayCommand]
         private async Task DeleteOrder(Order order)
         {
-            /*if (order != null)
+            if (order != null)
             {
                 Orders.Remove(order);
                 await _databaseService.DeleteOrderAsync(order);
-            }*/
-        }
 
+                OnPropertyChanged(nameof(Orders));
+            }
+        }
+        [RelayCommand]
+        private async Task OpenOrderModal(Order order) { 
+            IsEditingOrder = true;
+            SelectedOrder = order;
+        }
+        [RelayCommand]
+        private async Task EditOrder()
+        {
+            if (SelectedOrder == null) return;
+            await _databaseService.UpdateOrderAsync(SelectedOrder);
+            IsEditingOrder = false;
+            LoadData();
+        }
     }
 }
